Add a second internal field to weak ref proxies.
This is used to hold a pointer which is unique to the module, and is checked
as part of the process of determining whether a value is a weak ref, in
`isWeakRef()`.

This additional check will work assuming the lack of ill intent of other
native code, but it is susceptible to "gaming" by malicious native code. That
said, it is nonetheless an improvement in correctness over the old version of
the check, and all bets are off with native code anyway.